+++ agent/handlers/v1/response_test.go
@@ -23,6 +23,7 @@ import (
 	apieni "github.com/aws/amazon-ecs-agent/agent/api/eni"
 	apitask "github.com/aws/amazon-ecs-agent/agent/api/task"
 	apitaskstatus "github.com/aws/amazon-ecs-agent/agent/api/task/status"
+	docker "github.com/fsouza/go-dockerclient"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -33,6 +34,9 @@ const (
 	containerID    = "cid"
 	containerName  = "sleepy"
 	eniIPv4Address = "10.0.0.2"
+	volName        = "volume1"
+	volSource      = "/var/lib/volume1"
+	volDestination = "/volume"
 )
 
 func TestTaskResponse(t *testing.T) {
@@ -63,6 +67,13 @@ func TestTaskResponse(t *testing.T) {
 						"IPv4Addresses": []interface{}{"10.0.0.2"},
 					},
 				},
+				"Volumes": []interface{}{
+					map[string]interface{}{
+						"DockerName":  volName,
+						"Source":      volSource,
+						"Destination": volDestination,
+					},
+				},
 			},
 		},
 	}
@@ -88,6 +99,13 @@ func TestTaskResponse(t *testing.T) {
 			{
 				ContainerPort: 80,
 				Protocol:      apicontainer.TransportProtocolTCP,
+			},
+		},
+		VolumesUnsafe: []docker.Mount{
+			{
+				Name:        volName,
+				Source:      volSource,
+				Destination: volDestination,
 			},
 		},
 	}
@@ -130,6 +148,13 @@ func TestContainerResponse(t *testing.T) {
 				"IPv4Addresses": []interface{}{"10.0.0.2"},
 			},
 		},
+		"Volumes": []interface{}{
+			map[string]interface{}{
+				"DockerName":  volName,
+				"Source":      volSource,
+				"Destination": volDestination,
+			},
+		},
 	}
 
 	container := &apicontainer.Container{
@@ -138,6 +163,13 @@ func TestContainerResponse(t *testing.T) {
 			{
 				ContainerPort: 80,
 				Protocol:      apicontainer.TransportProtocolTCP,
+			},
+		},
+		VolumesUnsafe: []docker.Mount{
+			{
+				Name:        volName,
+				Source:      volSource,
+				Destination: volDestination,
 			},
 		},
 	}
@@ -189,4 +221,27 @@ func TestPortBindingsResponse(t *testing.T) {
 	assert.Equal(t, uint16(80), PortBindingsResponse[0].ContainerPort)
 	assert.Equal(t, uint16(80), PortBindingsResponse[0].HostPort)
 	assert.Equal(t, "tcp", PortBindingsResponse[0].Protocol)
+}
+
+func TestVolumesResponse(t *testing.T) {
+	container := &apicontainer.Container{
+		Name: containerName,
+		VolumesUnsafe: []docker.Mount{
+			{
+				Name:        volName,
+				Source:      volSource,
+				Destination: volDestination,
+			},
+		},
+	}
+
+	dockerContainer := &apicontainer.DockerContainer{
+		Container: container,
+	}
+
+	VolumesResponse := NewVolumesResponse(dockerContainer)
+
+	assert.Equal(t, volName, VolumesResponse[0].DockerName)
+	assert.Equal(t, volSource, VolumesResponse[0].Source)
+	assert.Equal(t, volDestination, VolumesResponse[0].Destination)
 }